Made pixels uint8

+++ dm_control/suite/wrappers/pixels.py
@@ -120,7 +120,7 @@ class Wrapper(dm_env.Environment):
       observation[STATE_KEY] = time_step.observation
 
     pixels = self._env.physics.render(**self._render_kwargs)
-    observation[self._observation_key] = pixels.astype('float32')
+    observation[self._observation_key] = pixels
 
     return time_step._replace(observation=observation)
